backend/requirements.txt
fastapi[standard]
haystack-ai
tika
motor
mongodb-atlas-haystack
sqlalchemy
psycopg2
nltk
transformers[torch,sentencepiece]
bs4
alembic
ollama-haystack
google-ai-haystack
pymilvus
milvus-haystack
pyjwt[crypto]
PyJWT
cryptography